ANDROID: sched: WALT: Add support for CFS_BANDWIDTH
cumulative runnable average is maintained in cfs_rq along with
rq so that when a cfs_rq is throttled/unthrottled, the contribution
of that cfs_rq can be updated at the rq level. Implement the
fixup_cumulative_runnable_avg callback for fair class to handle
the cfs_rq cumulative runnable average updates when the runnable
tasks demand is changed.

Bug: 139071966
Change-Id: Iccd473677cf491920aa82a6fc7e0a5374e5bb27f
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: sohamxda7 <sensoham135@gmail.com>